requirements.txt
fastapi
uvicorn
moviepy
openai-whisper
transformers
torch
numpy
scikit-learn
python-multipart
sentence-transformers
